requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
fastapi==0.115.8
pydantic-settings==2.8.0
uvicorn==0.34.0
